+++ portal/apps/homev3/tests/test_home.py
@@ -7,34 +7,40 @@ from django.test.client import Client
 from django.contrib.auth.models import User
 
 
-class SubscribeTestCase(TestCase):
+class HomeTestCase(TestCase):
 
     fixtures = ['test']
     http_host_header_param = {'HTTP_HOST': settings.SITE_DOMAIN}
-    # Commented urls are failing because there are problems loading the fixtures. TODO: fix
     urls_to_test = (
-        # {'url': '/'},
+        {'url': '/'},
         {'url': '/periodista/test-journalist/'},
         {'url': '/columnista/test-columnist/'},
-        # {'url': '/test/'},
+        {'url': '/test/'},
         {'url': '/tags/test/'},
         {'url': '/seccion/news/'},
-        # {'url': '/seccion/areasection/'},
+        {'url': '/seccion/areasection/'},
         {'url': '/articulo/2020/11/test-article/', 'headers': http_host_header_param},
         {'url': '/articulo/2020/11/test-article/', 'amp': True, 'headers': http_host_header_param},
-        # {'url': '/spinoff/'},
-        # {'url': '/spinoff/articulo/2020/11/test-article2/', 'headers': http_host_header_param},
-        # {'url': '/spinoff/articulo/2020/11/test-article2/', 'amp': True, 'headers': http_host_header_param},
+        {'url': '/spinoff/'},
+        {'url': '/spinoff/articulo/2020/11/test-article2/', 'headers': http_host_header_param},
+        {'url': '/spinoff/articulo/2020/11/test-article2/', 'amp': True, 'headers': http_host_header_param},
         {'url': '/test/articulo/2020/11/test-article3/', 'headers': http_host_header_param},
         {'url': '/test/articulo/2020/11/test-article3/', 'amp': True, 'headers': http_host_header_param},
     )
 
     def test_home(self):
         c = Client()
-        with self.settings(DEBUG=True):
+        with self.settings(DEBUG=True, DEFAULT_PUB="default"):
             for item in self.urls_to_test:
                 response = c.get(item['url'], {'display': 'amp'} if item.get('amp') else {}, **item.get('headers', {}))
                 self.assertEqual(response.status_code, 200, (response.status_code, response))
+            # status 200 also for the display param with a "not considered" value
+            item = self.urls_to_test[0]
+            response = c.get(item['url'], {'display': 'x'}, **item.get('headers', {}))
+            self.assertEqual(response.status_code, 200, (response.status_code, response))
+            # and with display=amp should return Forbidden
+            response = c.get(item['url'], {'display': 'amp'}, **item.get('headers', {}))
+            self.assertEqual(response.status_code, 403, (response.status_code, response))
 
     def test_home_logged_in(self):
         email, password = 'u1@example.com', User.objects.make_random_password()
@@ -43,7 +49,7 @@ class SubscribeTestCase(TestCase):
         user.save()
         c = Client()
         c.login(username=email, password=password)
-        with self.settings(DEBUG=True):
+        with self.settings(DEBUG=True, DEFAULT_PUB="default"):
             for item in self.urls_to_test:
                 response = c.get(item['url'], {'display': 'amp'} if item.get('amp') else {}, **item.get('headers', {}))
                 self.assertEqual(response.status_code, 200)
@@ -55,7 +61,7 @@ class SubscribeTestCase(TestCase):
         user.save()
         c = Client()
         c.login(username=email, password=password)
-        with self.settings(DEBUG=True):
+        with self.settings(DEBUG=True, DEFAULT_PUB="default"):
             for item in self.urls_to_test + ({'url': '/admin/'}, ):
                 response = c.get(item['url'], {'display': 'amp'} if item.get('amp') else {}, **item.get('headers', {}))
                 self.assertEqual(response.status_code, 200)