+++ portal/context_processors.py
@@ -60,10 +60,11 @@ def publications(request):
         slug_var = p.slug.replace('-', '_')
         result.update({slug_var.upper() + '_SUB': p.slug, slug_var + '_pub': p})
 
+    is_amp_detect = getattr(request, "is_amp_detect", False)
     result['extra_header_template'] = getattr(
-        settings, 'HOMEV3_EXTRA_HEADER_TEMPLATE%s' % ('_AMP' if request.is_amp_detect else ''), None
+        settings, 'HOMEV3_EXTRA_HEADER_TEMPLATE%s' % ('_AMP' if is_amp_detect else ''), None
     )
-    if not request.is_amp_detect:
+    if not is_amp_detect:
         result['footer_template'] = settings.HOMEV3_FOOTER_TEMPLATE
         result['subscribe_notice_template'] = getattr(
             settings, "HOMEV3_SUBSCRIBE_NOTICE_TEMPLATE", "homev3/templates/subscribe_notice.html"